frontend: Refactor apiProxy, move PortForward type inside api folder
This commit refactors apiProxy file into separate multiple files, also
improving types for them.

PortForward interface which was called PortForwardState was moved from
react component into portForward.ts, next to the api functions for port
forwarding logic
