+++ frontend/src/lib/k8s/apiProxy/apiProxy.test.ts
@@ -9,7 +9,7 @@ import WS from 'vitest-websocket-mock';
 import exportFunctions from '../../../helpers';
 import * as auth from '../../auth';
 import * as cluster from '../../cluster';
-import * as apiProxy from '.';
+import * as apiProxy from '../apiProxy';
 
 const baseApiUrl = exportFunctions.getAppUrl();
 const wsUrl = baseApiUrl.replace('http', 'ws');
